fix(checkout-service): fixed return types and handled input and responses checks (55-3936)

+++ checkout-service/src/main/java/com/team05/fooddelivery/checkout/service/PaymentService.java
@@ -1,7 +1,12 @@
 package com.team05.fooddelivery.checkout.service;
 import com.team05.fooddelivery.checkout.dto.RevenueReportDTO;
+import com.team05.fooddelivery.checkout.enums.OfferDiscountType;
 import com.team05.fooddelivery.checkout.enums.PaymentStatus;
+import com.team05.fooddelivery.checkout.model.Offer;
 import com.team05.fooddelivery.checkout.model.Payment;
+import com.team05.fooddelivery.checkout.model.PaymentOffer;
+import com.team05.fooddelivery.checkout.repository.OfferRepository;
+import com.team05.fooddelivery.checkout.repository.PaymentOfferRepository;
 import com.team05.fooddelivery.checkout.repository.PaymentRepository;
 import org.springframework.stereotype.Service;
 import com.team05.fooddelivery.checkout.dto.UserPaymentSummaryDTO;
@@ -18,9 +23,13 @@ import java.util.stream.Stream;
 public class PaymentService {
 
     private final PaymentRepository paymentRepository;
+    private final OfferRepository offerRepository;
+    private final PaymentOfferRepository paymentOfferRepository;
 
-    public PaymentService(PaymentRepository paymentRepository) {
+    public PaymentService(PaymentRepository paymentRepository,  OfferRepository offerRepository,  PaymentOfferRepository paymentOfferRepository) {
         this.paymentRepository =  paymentRepository;
+        this.offerRepository = offerRepository;
+        this.paymentOfferRepository = paymentOfferRepository;
     }
 
     // Payment CRUD
@@ -33,7 +42,7 @@ public class PaymentService {
     }
 
     public Payment getPaymentById(Long id) {
-        return paymentRepository.findById(id).orElseThrow(() -> new RuntimeException("Product not found"));
+        return paymentRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Payment not found"));
     }
 
     public Payment updatePayment(Long id, Payment updatedPayment) {
@@ -48,8 +57,12 @@ public class PaymentService {
     }
 
     public void deletePaymentById(Long id) {
+        if (!paymentRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Payment not found");
+        }
         paymentRepository.deleteById(id);
     }
+
     // S5-F1: Get Payments by Status and Date Range
     public List<Payment> getPaymentsByStatusAndDateRange(
             PaymentStatus status,
@@ -69,7 +82,7 @@ public class PaymentService {
         }
         Map<String, Object> transactionDetails = payment.getTransactionDetails();
         transactionDetails.put("refundReason", reason);
-        transactionDetails.put("refundAt", LocalDateTime.now().toString());
+        transactionDetails.put("refundedAt", LocalDateTime.now().toString());
         payment.setStatus(PaymentStatus.REFUNDED);
 
         return paymentRepository.save(payment);
@@ -105,7 +118,7 @@ public class PaymentService {
 
         return revenueReport;
     }
-      
+
     // S5-F7: Retry Failed Payment (Transactional)
     @Transactional
     public Payment retryFailedPayment(Long id) {
